fixed bugs in application.py

+++ application.py
@@ -102,6 +102,7 @@ def login():
 
         # Remember which user has logged in
         session["user_id"] = rows[0]["userID"]
+        session["logged_in"] = True
 
         # Redirect user to home page
         return redirect("/")
@@ -151,7 +152,9 @@ def checkout():
     # remove user from berg table (user is no longer in berg)
     # keep track of elapsed time, indicating how long the user spent in berg
     start_time_dictionary = db.execute("SELECT checkInTime FROM berg WHERE userID = :userID", userID=session["user_id"])
-    start_time = start_time_dictionary["checkInTime"][0]
+    if not start_time_dictionary:
+        return apology("user isn't checked in")
+    start_time = start_time_dictionary[0]["checkInTime"]
     db.execute("DELETE FROM berg WHERE userID = :userID", userID=session["user_id"])
     end_time = datetime.now(timezone('US/Eastern')).time()
     elapsed_time = end_time - start_time
